Protocol: Version management.
This and the version hash allows for an at-a-glance understanding
of whether or not two protocols are the same, and allows for a
simple version tag to be used in discussing the protocol.

Doesn't help in the world of branching changes, but for the
basic evolution of any given protocol, it should provide an
understanding of how it differs from previous versions.

Semantic versioning is baked in.

Minor revisions are automatically bumped.

New functionality can be demarked by manually bumping the
"feature" release.

Any "feature" level version of a protocol should be able to run
on the same deck layout as previous versions.

A breaking or major change (ie, entirely new deck layout,
additional instruments, new modules), can be bumped as a "major"
release.