Protocol: Command refactor.
Holy code reproduction, Batman!

This is a pretty rushed commit; I need to be done with this
project, so I don't have much time.

If I had some more bandwidth or a do-over, I'd probably make a
Command class that handles all this complexity.

There's also inconsistent handling of touchtip and blowout because
they're not really used in MotorHandler yet.  I just added them to
transfer for backwards compatibility reasons and to stub them out
for later expansion.