Tipracks: Individual tip inventories.
Not nearly as performant, but we're not optimizing for that right now.

This is going to let us do multichannel.

I'd meant to change get_next_tip to get_clean_tip so that it's not a destructive
operation, and have the callers mark the tip as used as appropriate, but I
didn't want to deal with all the failing tests.

This commit was already quite the challenge.