ItemGroup equality.
For such a simple method, I'm still not sure I've thought through all the scenarios enough.

Debated for a while on whether ItemGroup([Item(1)]) should equal Item(1), so long as Item
is implemented in a way that makes Item(1) == Item(1).  Even did the implementation, but
then I realized that even if ItemGroup([Item(1)]) was smart enough to check equality with
the first element in the Item, every class inside an ItemGroup would also have to be aware
of ItemGrouop to properly do equalty.  Since Item(1) != ItemGroup([Item(1)]) unless I also
overload Item's __eq__ and specifically account for it...

Programming is hard.