ContextHandler passes filter kwargs to boolean methods.
This gets rid of all the complicated stuff around volume searching and support more dynamic queries
on methods that take properties.

I also renamed volume to has_volume because volume as a method that acts like a property but takes
an argument and returns a boolean instead of the volume of the instrument is gross.