Filter: filter_objects as a generalized approach.
Of course, it doesn't work for ContextHandler because the
handler needs the axis key from the _instruments object for some
reason.

I hate everything about instrument axises work; there's some
real messed up stuff in the implementation that keeps causing
strange errors and it's so interconnected that it's super hard
to refactor.