Protocol: Clean up calibration and instrument storage.
Now the Context Handler bubbles up its calibration query to the
parent Protocol and the Pipette bubbles up its calibration query
to the parent Context, which mean the Protocol is the canonical
store of the calibration data.

Also realized that the reason _instruments works funny is because
Protocol has an _instruments which is a mapping of axises and the
tool name, unlike Context, which has a mapping of the axises and
the instrument instances.

So I changed Protocol._instruments to Protocol._head.

This is already much cleaner!