Protocol: Get calibration requirements.
This returns an object containing all of the required calibration
points for a particular Protocol so that the user interface can
prompt for these things to happen.

Needs much more thorough tests, but without a machine to run this
on, I can't really do much with transfer_group, consolidate,
distribute or mix besides ensuring they run without errors.

I really should have just stuck to transfer for the example
implementation, but really, with only one command, it would have
been hard to demonstrate that multiple are possible.

If I ever get a big chunk more time on this, I'll be redoing the
command architecture anyways.

The whole Handlers concept needs a bit of tweaking, too, at least
when running them as an isolated test case like we're doing here.