Test coverage runthrough.
Still not 100%, but that's mostly IO stuff and the intricacies of
the custom SmoothieBoard firmware my company was using at the
time, which is not worth mocking up since it's already been
deprecated.

Protocol could use a tad bit more coverage, but that's best to be
left for another day.